metrics_cache::cli_args

Struct CliArgs

pub struct CliArgs {

    pub namespace: Option<String>,
    pub metrics_fetcher_daemonset_name: Option<String>,
    pub ingest_address: String,
    pub ingest_port: u16,
    pub ingest_service_name: Option<String>,
    pub ingest_tls_secret: Option<String>,
    pub ingest_tls_secret_generate_if_missing: bool,
    pub ingest_tls_secret_generation_validity: Duration,
    pub reader_allowlist: Vec<String>,
    pub writer_allowlist: Vec<String>,
    pub kubelet_stats_cache_ttl: Duration,
    pub system_agent_cache_ttl: Duration,
    pub kubelet_health_cache_ttl: Duration,
    pub log_level: String,
    pub pull_address: String,
    pub pull_port: u16,
    pub pull_service_name: Option<String>,
    pub pull_tls_secret: Option<String>,
    pub pull_tls_secret_generate_if_missing: bool,
    pub pull_tls_secret_generation_validity: Duration,
    pub max_retries: u8,
    pub connect_timeout: Duration,
    pub read_timeout: Duration,
    pub cluster_name: String,
    pub cluster_host_name: String,
    pub import_all_annotations: bool,
    pub annotation_key_pattern: Option<Regex>,
    pub disable_pull_authentication: bool,
    pub pull_shared_secret: Option<String>,
    pub push_interval: Duration,
    pub push_certificate_renewal_threshold: Duration,
    pub push_receiver: Option<String>,
    pub push_registration_ott: Option<String>,
    pub push_registration_pem: Option<String>,
    pub push_registration_insecure_skip_site_ca_verification: bool,
    pub excluded_node_role_patterns: Vec<Regex>,
    pub namespace_include_patterns: Vec<Regex>,
    pub namespace_exclude_patterns: Vec<Regex>,
    pub otel_endpoint: Option<String>,
    pub otel_push_interval: Duration,
    pub api_health_poll_interval: Duration,
    pub all_pods: bool,
    pub include_cronjob_pods: bool,
    pub all_namespaces: bool,
    pub all_nodes: bool,
    pub all_deployments: bool,
    pub all_daemonsets: bool,
    pub all_statefulsets: bool,
    pub all_cronjobs: bool,
    pub all_pvcs: bool,

}

Fields
namespace: Option<String>

The namespace metrics-cache is running in

metrics_fetcher_daemonset_name: Option<String>

Name of the metrics-fetcher DaemonSet whose scheduled Pods are expected to report metrics

ingest_address: String

IP address to bind to for intra-cluster ingest

ingest_port: u16

Port to bind to for intra-cluster ingest

ingest_service_name: Option<String>

The service name for intra-cluster ingest

ingest_tls_secret: Option<String>

Secret containing the TLS certificate and key for intra-cluster ingest

ingest_tls_secret_generate_if_missing: bool

Generate the intra-cluster ingest TLS CA and certificate if missing and store it as a Kubernetes Secret

ingest_tls_secret_generation_validity: Duration

When generating the intra-cluster ingest TLS CA and certificate, specifies how long they should be valid for in days

reader_allowlist: Vec<String>

Service accounts that have access to query data from the metrics cache API GET endpoints. Comma-separated, in the form NAMESPACE:SERVICEACCOUNT

writer_allowlist: Vec<String>

Service accounts that have access to query data from the metrics cache API POST endpoints. Comma-separated, in the form NAMESPACE:SERVICEACCOUNT

kubelet_stats_cache_ttl: Duration

How long (seconds) kubelet stats entries are persisted in the cache

system_agent_cache_ttl: Duration

How long (seconds) system agent entries are persisted in the cache

kubelet_health_cache_ttl: Duration

How long (seconds) kubelet health entries are persisted in the cache

log_level: String

How verbose to log

pull_address: String

IP address to bind to for pull mode

pull_port: u16

Port to bind to for pull mode

pull_service_name: Option<String>

The service name for pull mode

pull_tls_secret: Option<String>

Secret containing the TLS certificate and key for pull mode

pull_tls_secret_generate_if_missing: bool

Generate the pull mode TLS CA and certificate if missing and store it as a Kubernetes Secret

pull_tls_secret_generation_validity: Duration

When generating the pull-mode TLS CA and certificate, specifies how long they should be valid for in days

max_retries: u8

Maximum number of retries when contacting the Kubernetes API for authentication

connect_timeout: Duration

Time in seconds to wait for a TCP connection to the Kubernetes API during authentication

read_timeout: Duration

Time in seconds to wait for a response from the Kubernetes API during authentication

cluster_name: String

The name of the cluster, included in each piggyback hostname

cluster_host_name: String

The name of the source host in Checkmk (should be an exact match)

import_all_annotations: bool

Import all annotations as host labels

annotation_key_pattern: Option<Regex>disable_pull_authentication: bool

Make the pull-based endpoints publicly accessible

pull_shared_secret: Option<String>

For pull mode, the shared secret that is also stored in Checkmk’s password store. If not set, pull mode is disabled and requests will give a 401. To disable auth, use –disable-pull-authentication

push_interval: Duration

Push interval in seconds for push mode. Ignored if push mode is not enabled.

push_certificate_renewal_threshold: Duration

Number of days before client certificate expiration to attempt renewal

push_receiver: Option<String>

Enable push mode and send sections to the specified server (including port)

push_registration_ott: Option<String>

Token to register with the Checkmk push-agent receiver for push mode

push_registration_pem: Option<String>

CA certificate (PEM) of the Checkmk site, used to verify we are about to register with the correct push-agent receiver

push_registration_insecure_skip_site_ca_verification: bool

Avoid verifying the identity of the configured push-agent receiver during initial registration. Do NOT use in production.

excluded_node_role_patterns: Vec<Regex>

Excluded node role (infix) pattern for cluster-level aggregations. May be specified multiple times

namespace_include_patterns: Vec<Regex>

Only emit piggyback hosts belonging to namespaces matching at least one of these regex patterns. May be specified multiple times

namespace_exclude_patterns: Vec<Regex>

Do not emit piggyback hosts belonging to namespaces matching any of these regex patterns. May be specified multiple times

otel_endpoint: Option<String>

Enable sending OTel metrics to the endpoint given

otel_push_interval: Duration

Push interval in seconds for sending otel metrics.

api_health_poll_interval: Duration

How often (seconds) to query the Kubernetes API health endpoints /readyz and /livez

all_pods: bool

Emit all Pod resources rather than only annotated ones

include_cronjob_pods: bool

Allow Pods owned by CronJobs to become standalone piggyback hosts

all_namespaces: bool

Emit all Namespace resources rather than only annotated ones

all_nodes: bool

Emit all Node resources rather than only annotated ones

all_deployments: bool

Emit all Deployment resources rather than only annotated ones

all_daemonsets: bool

Emit all DaemonSet resources rather than only annotated ones

all_statefulsets: bool

Emit all StatefulSet resources rather than only annotated ones

all_cronjobs: bool

Emit all CronJob resources rather than only annotated ones

all_pvcs: bool

Emit PVC sections on Pod and workload resources

Implementations

impl CliArgs

pub fn ingest_tls_config(&self) -> TlsConfig

pub fn pull_tls_config(&self) -> TlsConfig