metrics_cache

Module host_settings

Structs

AlwaysEmitted

Holds information (based on CLI arguments) about which resources should always be emitted. The ones which are false here will only be emitted if explicitly promoted to piggyback host via an annotation.

HostSettings

NamespaceFilter

Controls which namespaces produce namespaced piggyback hosts.

Enums

AnnotationKeyPattern

Functions

node_roles

The roles of a node, derived from its node-role.kubernetes.io/<role> labels. A node can carry any number of roles, including none at all.