metrics_cache::ingest

Module reflectors

Macros

define_reflectors 🔒

Structs

FrozenReflectorHealths 🔒

FrozenStores

ReflectorHealth 🔒

The inner-state of a reflector. This gets updated by the reflector’s inspect callback as certain events happen (namely Init, InitDone, and errors).

ReflectorHealthHandle 🔒

ReflectorHealthHandles 🔒

The collection of ReflectorHealthHandles which are constantly being changed by the reflector as events happen. Thus, we can not rely on this for a snapshot; at snapshot-time, it is only used to generate a FrozenReflectorHealths via Self::freeze().

Stores

The [Store]s where all of the watched Kubernetes kinds that rustik monitors end up. The stores here are not frozen and might update in the background at any time.

Constants

REFLECTOR_COUNT 🔒

Functions

start_reflector 🔒