metrics_cache::piggyback

Trait PiggybackHost

pub(crate) trait PiggybackHost {
    // Required methods
    fn metadata(&self) -> Option<&ObjectMeta>;
    fn kind(&self) -> &str;
    fn emit(&self) -> Vec<Result<WriteableSection, SectionError>>;

    // Provided methods
    fn namespace_for_filtering(&self) -> Option<&str> { ... }
    fn annotation_emit_override(&self) -> Option<bool> { ... }
}

Represents a piggyback host for which to emit/write a section data.

Required Methods

fn metadata(&self) -> Option<&ObjectMeta>

fn kind(&self) -> &str

fn emit(&self) -> Vec<Result<WriteableSection, SectionError>>

Provided Methods

fn namespace_for_filtering(&self) -> Option<&str>

The namespace used to decide whether this host is filtered. Cluster resources return None and are unaffected by namespace filtering.

fn annotation_emit_override(&self) -> Option<bool>

Answers the question: Does an annotation request this resource become a piggyback host?

With an annotation like:

annotations:
  checkmk.com/promote-to-host: "true"

in a resource’s mewtadata, we will promote it to become a piggyback host in Checkmk (assuming we support monitoring the resource).

There are also CLI flags (set via helm chart variables) that control at the kind-level which resources get promoted. We do not check that here but rather in crate::piggyback::collect() / crate::piggyback::should_emit(). A valid true/false annotation overrides the global setting in all cases, otherwise the kind’s global flag decides.

If the value is “true” or “false, we return the boolean of it, otherwise we return None and log a warning with identifying info.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".