metrics_cache::push

Function push_loop

pub async fn push_loop(
    client: CheckmkPushClient,
    registration: CheckmkPushRegistration<'_>,
    state: AppState<impl TokenValidator>,
    push_interval: Duration,
    certificate_renewal_threshold: Duration,
) -> Result<(), WriterDropped>

Run the push loop forever.

Every push interval, push_cycle() is called to generate sections and send them to the Checkmk server.

Waits until all stores are ready before doing its initial push and looping.

After each push attempt, renews the client certificate when needed.