metrics_cache::section

Module self_health

Self-health section(s) for rustik.

The sections here provide information about the current health of rustik and its two components (metrics-cache and metrics-fetcher).

These get emitted on the cluster piggyback host.

Structs

KubeRustikHealthV1 🔒

MetricsCacheMetadata 🔒

MetricsFetcherIngestionHealth 🔒

NodeMetricsFetcherHealth 🔒

ReflectorHealth 🔒

Functions

duration_to_secs 🔒

By default serde serializes Duration as a dict of time units (mins, secs, etc.). This is less useful for parsing later on, we want pure seconds, so this is used to teach serde how to serialize a cache map from the snapshot’s SelfHealth in seconds.