metrics_cache

Module snapshot

Modules

indexes

kubelet_health

metric_tables

owner_graph

self_health

system_agent

Structs

Snapshot

Represents a single, static snapshot of the state of the cluster as it pertains to Checkmk monitoring.

Uid

The unique ID of a given Kubernetes object.