metrics_cache::startup::tls

Function get_or_create_cert_secret

async fn get_or_create_cert_secret(
    secrets: &Api<Secret>,
    secret_name: &str,
    namespace: &str,
    service_name: &str,
    longevity: Duration,
) -> Result<Secret>

First, attempt to find a matching Secret in the current namespace. If one is found, return it immediately. Otherwise attempt to create a Secret and return the new one.

Has the semantics of create_cert_secret() during creation: If during creation we find that a matching secret has been created, we will read it and return that (preventing potential races).

If any other error occurs that is not a “not found” error, we return it immediately.